// isoft-fe/src/app/models/achievement.model.ts
export interface Achievement {
    code: string;
    createAt: string;
    date: Date | string;
    deleteAt: Date | string;
    description: string;
    filePath: string;
    fullName: string;
    id: number;
    isDelete: boolean;
    name: string;
    note: string;
    updateAt: Date | string;
    userId: number;
}